1.更改体重，将体重设置为浮点型

+++ src/main/java/org/cqu/edu/msc/annihilation/campephilus/module/core/domain/info/PatientInfo.java
@@ -69,7 +69,7 @@ public class PatientInfo extends BaseInfoSuperclass implements Serializable {
      */
     @NotNull(message = "weight must cannot empty")
     @Column(name = "weight")
-    private Integer weight;
+    private Float weight;
 
     /**
      * 年龄
